AI-Voice-Detection-main/requirements.txt
# Core dependencies
setuptools>=65.0.0
wheel

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Providers
groq>=0.4.0
openai>=1.10.0

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# HTTP client (for testing and AssemblyAI)
requests>=2.31.0

# Database (Supabase)
supabase>=2.0.0

# Async Task Queue (Celery)
celery>=5.3.0
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Optional: Legacy voice detection
# librosa==0.10.1
# numpy>=1.24.0,<2.0.0
# joblib==1.3.2
# scikit-learn==1.3.2
# soundfile==0.12.1
